clean: Move *Runner* from common.markdown.exec to common.exec
