+++ java/dev/enola/chat/BUILD
@@ -45,7 +45,6 @@ java_library(
         "//java/dev/enola/common/exec",
         "//java/dev/enola/common/io",
         "//java/dev/enola/common/linereader",
-        "//java/dev/enola/common/markdown/exec:markdown",
         "//java/dev/enola/common/secret",
         "//java/dev/enola/data/id",
         "//java/dev/enola/identity",
@@ -68,6 +67,7 @@ junit_tests(
         ":chat",
         "//java/dev/enola/common/context",
         "//java/dev/enola/common/context/testlib",
+        "//java/dev/enola/common/exec",
         "//java/dev/enola/common/markdown/exec:markdown",
         "//java/dev/enola/common/secret",
         "//java/dev/enola/identity",
