+++ java/dev/enola/common/markdown/exec/ExecMD.java
@@ -21,6 +21,9 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 
 import com.google.common.io.Files;
 
+import dev.enola.common.exec.Runner;
+import dev.enola.common.exec.VorburgerExecRunner;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
